Removed old classes and fixes tests.
Fixing test required that the addProperty(String, Class<P>) method is removed.
It caused problems where users would call addProperty("int", Integer.class) and
then register listener onChange(MyClass::updateInt), where updateInt takes "int"
instead of "Integer". The original property would be set to null and that value
would be passed to the listener, breaking it.

This problem still persists if the client-side send a bad packet with misaligned
data - ticket created.
