--- flow-server/src/main/java/com/vaadin/flow/component/webcomponent/WebComponentMethod.java
@@ -1,47 +1,0 @@
-/*
- * Copyright 2000-2018 Vaadin Ltd.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- */
-package com.vaadin.flow.component.webcomponent;
-
-import java.lang.annotation.Documented;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-/**
- * Marks a method to be published and synced on client for the generated server
- * side WebComponent.
- */
-@Retention(RetentionPolicy.RUNTIME)
-@Target({ ElementType.METHOD })
-@Documented
-public @interface WebComponentMethod {
-
-    /**
-     * Name of the property.
-     *
-     * @return published property name
-     */
-    String value();
-
-    /**
-     * Initial value for the property if nothing is defined on the client side.
-     *
-     * @return property inital value
-     */
-    String initialValue() default "";
-
-}
