+++ flow-server/src/main/java/com/vaadin/flow/server/startup/WebComponentRegistryInitializer.java
@@ -19,153 +19,140 @@ import javax.servlet.ServletContainerInitializer;
 import javax.servlet.ServletContext;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.HandlesTypes;
-import java.lang.reflect.Field;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Map;
 import java.util.Set;
-import java.util.function.Function;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 import com.vaadin.flow.component.Component;
-import com.vaadin.flow.component.WebComponent;
-import com.vaadin.flow.component.webcomponent.WebComponentMethod;
-import com.vaadin.flow.component.webcomponent.WebComponentProperty;
+import com.vaadin.flow.component.WebComponentExporter;
 import com.vaadin.flow.internal.CustomElementNameValidator;
 import com.vaadin.flow.server.InvalidCustomElementNameException;
-import com.vaadin.flow.server.webcomponent.WebComponentRegistry;
+import com.vaadin.flow.server.webcomponent.WebComponentBuilder;
+import com.vaadin.flow.server.webcomponent.WebComponentBuilderRegistry;
 
 /**
- * Servlet initializer for collecting all available {@link WebComponent}
- * annotated classes on startup.
+ *
+ * Servlet initializer for collecting all classes that extend
+ * {@link WebComponentExporter} on startup.
+ *
+ * @author Vaadin Ltd.
+ * @since
  */
-@HandlesTypes({ WebComponent.class })
+@HandlesTypes({ WebComponentExporter.class })
 public class WebComponentRegistryInitializer
         implements ServletContainerInitializer {
 
     @Override
     public void onStartup(Set<Class<?>> set, ServletContext servletContext)
             throws ServletException {
-        WebComponentRegistry instance = WebComponentRegistry
+        WebComponentBuilderRegistry instance = WebComponentBuilderRegistry
                 .getInstance(servletContext);
         if (set == null || set.isEmpty()) {
-            instance.setWebComponents(Collections.emptyMap());
+            instance.setWebComponentBuilders(Collections.emptySet());
             return;
         }
-        Set<? extends Class<? extends Component>> componentSet = set.stream()
-                .filter(Component.class::isAssignableFrom)
-                .filter(webComponent -> webComponent
-                        .isAnnotationPresent(WebComponent.class))
-                .map(clazz -> (Class<? extends Component>) clazz)
+        Set<Class<?>> exporterClasses = set.stream()
+                .filter(WebComponentExporter.class::isAssignableFrom)
+                .filter(c -> !c.isInterface())
                 .collect(Collectors.toSet());
+        Set<WebComponentExporter<? extends Component>> exporters;
 
-        validateDistinct(componentSet);
-        validateComponentName(componentSet);
+        exporters = constructExporters(exporterClasses);
+        validateDistinct(exporters);
+        validateComponentName(exporters);
 
-        componentSet.forEach(this::validateMethodsAndProperties);
+        Set<WebComponentBuilder<? extends Component>> builders =
+                constructBuilders(exporters);
 
-        Map<String, Class<? extends Component>> webComponentMap = componentSet
-                .stream().collect(Collectors
-                        .toMap(this::getWebComponentName, Function.identity()));
+        instance.setWebComponentBuilders(builders);
+    }
 
-        instance.setWebComponents(webComponentMap);
+    private Set<WebComponentBuilder<? extends Component>> constructBuilders(Set<WebComponentExporter<? extends Component>> exporters) {
+        return exporters.stream()
+                .map(exporter -> {
+                    try {
+                        return new WebComponentBuilder<>(exporter);
+                    } catch (IllegalArgumentException e) {
+                        throw new IllegalArgumentException(String.format(
+                                "Invalid exporter '%s': %s",
+                                exporter.getClass().getCanonicalName(),
+                                e.getMessage()));
+                    }
+                }).collect(Collectors.toSet());
+    }
+
+    private Set<WebComponentExporter<? extends Component>> constructExporters(
+            Set<Class<?>> exporterClasses) {
+
+        //Instantiator instantiator = VaadinService.getCurrent()
+        // .getInstantiator();
+
+        Set<WebComponentExporter<? extends Component>> set = new HashSet<>();
+        WebComponentExporter<? extends Component> webComponentExporter;
+        for (Class<?> exporterClass : exporterClasses) {
+            try {
+                webComponentExporter=
+                    (WebComponentExporter<? extends Component>) exporterClass.newInstance();
+            } catch (InstantiationException | IllegalAccessException e) {
+                throw new RuntimeException(String.format("Could not construct %s " +
+                        "from %s!", WebComponentExporter.class.getName(),
+                        exporterClass.getCanonicalName()), e);
+            }
+            set.add(webComponentExporter);
+        }
+
+        return set;
     }
 
     /**
      * Validate that all web component names are valid custom element names.
      *
-     * @param componentSet
+     * @param exporterSet
      *         set of web components to validate
      */
     protected void validateComponentName(
-            Set<? extends Class<? extends Component>> componentSet) {
-        for (Class<? extends Component> clazz : componentSet) {
-            String tagName = getWebComponentName(clazz);
-            if (!CustomElementNameValidator.isCustomElementName(tagName)) {
+            Set<WebComponentExporter<? extends Component>> exporterSet) {
+        for (WebComponentExporter<? extends Component> exporter : exporterSet) {
+            String tag = exporter.tag();
+            if (!CustomElementNameValidator.isCustomElementName(tag)) {
                 String msg = String
-                        .format("WebComponent name '%s' for '%s' is not a valid custom element name.",
-                                tagName, clazz.getCanonicalName());
+                        .format("Tag name '%s' given by '%s' is not a valid " +
+                                        "custom element name.",
+                                tag, exporter.getClass().getCanonicalName());
                 throw new InvalidCustomElementNameException(msg);
             }
         }
     }
 
     /**
-     * Validate that in all the components we only have one instance for each
-     * element name.
+     * Validate that we have exactly one web component exporter per tag name.
      *
-     * @param componentSet
+     * @param exporterSet
      *         set of web components to validate
      */
     protected void validateDistinct(
-            Set<? extends Class<? extends Component>> componentSet) {
-        long count = componentSet.stream().map(this::getWebComponentName)
+            Set<WebComponentExporter<? extends Component>> exporterSet) {
+        long count = exporterSet.stream().map(WebComponentExporter::tag)
                 .distinct().count();
-        if (componentSet.size() != count) {
+        if (exporterSet.size() != count) {
             Map<String, Class<?>> items = new HashMap<>();
-            for (Class<? extends Component> component : componentSet) {
-                String webComponentName = getWebComponentName(component);
-                if (items.containsKey(webComponentName)) {
-                    String message = String
-                            .format("Found two WebComponents with the same name for classes '%s' and '%s'",
-                                    items.get(webComponentName).getName(),
-                                    component.getName());
+            for (WebComponentExporter<? extends Component> exporter : exporterSet) {
+                String tag = exporter.tag();
+                if (items.containsKey(tag)) {
+                    String message = String.format(
+                            "Found two %s classes '%s' and '%s' for the tag " +
+                                    "name '%s'. Tag must be unique.",
+                            WebComponentExporter.class.getSimpleName(),
+                            items.get(tag).getCanonicalName(),
+                            exporter.getClass().getCanonicalName(),
+                            tag);
                     throw new IllegalArgumentException(message);
                 }
-                items.put(webComponentName, component);
+                items.put(tag, exporter.getClass());
             }
         }
-    }
-
-    protected void validateMethodsAndProperties(Class<?> webComponent) {
-        Set<String> methods = getWebComponentMethods(webComponent);
-        Set<String> propertyFields = getWebComponentPropertyFields(
-                webComponent);
-
-        Set<String> duplicates = methods.stream()
-                .filter(propertyFields::contains).collect(Collectors.toSet());
-
-        if (!duplicates.isEmpty()) {
-            String message = String
-                    .format("In the WebComponent '%s' there is a method and a property for the name(s) %s",
-                            webComponent, duplicates);
-            throw new IllegalArgumentException(message);
-        }
-    }
-
-    private Set<String> getWebComponentPropertyFields(Class<?> webComponent) {
-        Set<String> propertyFields = new HashSet<>();
-
-        if (webComponent.getSuperclass() != null) {
-            propertyFields.addAll(getWebComponentPropertyFields(
-                    webComponent.getSuperclass()));
-        }
-        Stream.of(webComponent.getDeclaredFields())
-                .filter(field -> WebComponentProperty.class
-                        .isAssignableFrom(field.getType())).map(Field::getName)
-                .forEach(propertyFields::add);
-
-        return propertyFields;
-    }
-
-    private Set<String> getWebComponentMethods(Class<?> webComponent) {
-        Set<String> methods = new HashSet<>();
-
-        if (webComponent.getSuperclass() != null) {
-            methods.addAll(
-                    getWebComponentMethods(webComponent.getSuperclass()));
-        }
-        Stream.of(webComponent.getDeclaredMethods()).filter(method -> method
-                .isAnnotationPresent(WebComponentMethod.class))
-                .map(method -> method.getAnnotation(WebComponentMethod.class))
-                .forEach(annotation -> methods.add(annotation.value()));
-
-        return methods;
-    }
-
-    protected String getWebComponentName(Class<? extends Component> clazz) {
-        return clazz.getAnnotation(WebComponent.class).value();
     }
 }
