+++ flow-server/src/test/java/com/vaadin/flow/server/communication/WebComponentProviderTest.java
@@ -1,11 +1,26 @@
-package com.vaadin.flow.server.communication;
+/*
+ * Copyright 2000-2018 Vaadin Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
 
-import java.io.ByteArrayOutputStream;
-import java.io.IOException;
-import java.util.Collections;
+package com.vaadin.flow.server.communication;
 
 import javax.servlet.ServletContext;
 import javax.servlet.http.HttpServletResponse;
+import java.io.ByteArrayOutputStream;
+import java.io.IOException;
+import java.util.HashSet;
 
 import org.junit.Assert;
 import org.junit.Before;
@@ -15,14 +30,17 @@ import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
 
 import com.vaadin.flow.component.Component;
-import com.vaadin.flow.component.WebComponent;
+import com.vaadin.flow.component.Tag;
+import com.vaadin.flow.component.WebComponentExporter;
+import com.vaadin.flow.component.webcomponent.WebComponentDefinition;
 import com.vaadin.flow.server.DefaultDeploymentConfiguration;
 import com.vaadin.flow.server.MockInstantiator;
 import com.vaadin.flow.server.VaadinResponse;
 import com.vaadin.flow.server.VaadinService;
 import com.vaadin.flow.server.VaadinServletRequest;
 import com.vaadin.flow.server.VaadinSession;
-import com.vaadin.flow.server.webcomponent.WebComponentRegistry;
+import com.vaadin.flow.server.webcomponent.WebComponentBuilder;
+import com.vaadin.flow.server.webcomponent.WebComponentBuilderRegistry;
 
 public class WebComponentProviderTest {
 
@@ -98,12 +116,13 @@ public class WebComponentProviderTest {
 
         Mockito.when(request.getServletContext()).thenReturn(servletContext);
         Mockito.when(request.getContextPath()).thenReturn("");
-        WebComponentRegistry registry = WebComponentRegistry
+        WebComponentBuilderRegistry registry = WebComponentBuilderRegistry
                 .getInstance(servletContext);
-        registry.setWebComponents(
-                Collections.singletonMap("my-component", MyComponent.class));
+        final HashSet<WebComponentBuilder<?>> set = new HashSet<>();
+        set.add(new WebComponentBuilder<>(new MyComponentExporter()));
+        registry.setWebComponentBuilders(set);
         Mockito.when(servletContext
-                .getAttribute(WebComponentRegistry.class.getName()))
+                .getAttribute(WebComponentBuilderRegistry.class.getName()))
                 .thenReturn(registry);
 
         ByteArrayOutputStream out = Mockito.mock(ByteArrayOutputStream.class);
@@ -125,7 +144,21 @@ public class WebComponentProviderTest {
 
     }
 
-    @WebComponent("my-component")
-    public class MyComponent extends Component {
+    @Tag("my-component")
+    public static class MyComponent extends Component {
     }
+
+    public static class MyComponentExporter implements WebComponentExporter<MyComponent> {
+
+        @Override
+        public String tag() {
+            return "my-component";
+        }
+
+        @Override
+        public void define(WebComponentDefinition<MyComponent> definition) {
+
+        }
+    }
+
 }
